feat(PullDownRefresh): use new style
