+++ src/pull-down-refresh/style/index.js
@@ -1,1 +1,1 @@
-import '../../_common/style/mobile/components/pull-down-refresh/_index.less';
+import '../../_common/style/mobile/components/pull-down-refresh/v2/_index.less';
